+++ CHANGELOG.md
@@ -1,16 +1,16 @@
 # Changelog and release notes
 
-## Unreleased
+## v0.11.1
 ### Features
 - add support for returning null instead of throwing authorization error (`authMode` property of `buildSchema` config)
 - add support for generating object type field in schema from method with `@FieldResolver`
+### Fixes
+- fix bug when converting object scalars to target class instance (#65)
 
 ## v0.11.0
 ### Features
 - add support for creating and attaching middlewares, guards and interceptors to fields and resolvers
 - **Breaking Change**: remove deprecated decorators with `GraphQL` prefix and `{ array: true }` type option
-### Fixes
-- fix bug when converting object scalars to target class instance (#65)
 
 ## v0.10.0
 ### Features